Handwriting_recognition_number_train_LeNet5/requirements.txt
keras==3.3.3
Pillow==10.3.0
PyQt5==5.15.10
PyQt5_sip==12.13.0
tensorflow==2.16.1
tensorflow_intel==2.16.1
